Use TypeToken, split out Manager/PluginProvidedService
I started going down the rabbit hole of type reification, and while neat, it's also a semi-solved problem. I don't intend to write a whole separate library for reification to my exact specifications, this will do.
As IntelliJ IDEA does not understand that the UnloadedService is to be removed only after compilation completes (it removes it out of order as soon as the class itself finishes compiling, it seems?) the UnloadedServiceTest is now a separate class so that it can easily be excluded. Maven has no such issues.