Extract digit-only comparison, return StringBuilder
Returning a StringBuilder instead of a finalized String has no impact on comparison and makes a few relatively serious optimizations. This yields an overall performance increase of about 20-25%, which is further increased by a large (albeit less consistent) factor during warmups or normal non-repeat use. In addition to the cost of string creation, AbstractStringBuilder#length is not recalculated with each call.